requirements.txt
crewai
langchain
together
python-dotenv